Add rough prototype of the program
The program has data baked in as global variables and it just plots
distance as a function of time.

The variables px10, py10, pz10 are the initial position of ship 1.
The variables px20, py20, pz20 are the initial position of ship 2.

The variables vx1, vy1, vz1 are the velocity components of ship 1.
The variables vx2, vy2, vz2 are the velocity components of ship 2.